models/example_test.go
package models

import (
	"testing"
	"time"
)

type MockExampleTable struct {
	GetFunc func(key string) (Example, error)
}

func (m *MockExampleTable) Get(userID string) (Example, error) {
	return m.GetFunc(userID)
}

func TestGetExample(t *testing.T) {
	// Create a mock implementation of ExampleRepo
	mockRepo := &baseExampleRepo{
		table: &MockExampleTable{
			GetFunc: func(key string) (Example, error) {
				// Create a sample example for testing
				example := Example{
					UserID:    key,
					CreatedAt: int(time.Now().Unix()),
					FieldName: 123,
				}
				return example, nil
			},
		},
	}

	// Call the Get function with a sample user ID
	userID := "testUser"
	example, err := mockRepo.Get(userID)
	if err != nil {
		t.Errorf("unexpected error: %v", err)
	}

	// Verify the returned example's user ID
	if example.UserID != userID {
		t.Errorf("expected user ID %s, got %s", userID, example.UserID)
	}

	// Verify any other assertions you have for the returned example

	// Add more test cases as needed
}
